backend/engine/requirements.txt
psycopg2-binary
pandas
geopandas
shapely
tensorflow
scikit-learn